447 - Moved method textWidth() from xssUtils to stringUtils http://localhost:4000/components/datagrid/example-tooltips.html
